+++ aspnet/RVTR.Media.Domain/Abstracts/AEntity.cs
@@ -5,6 +5,6 @@ namespace RVTR.Media.Domain.Abstracts
   /// </summary>
   public abstract class AEntity
   {
-    public int EntityId { get; set; }
+    public long EntityId { get; set; }
   }
 }
